@@ -3,9 +3,12 @@
 #
 cmake_minimum_required(VERSION 3.14)
 
-find_package(Torch REQUIRED)
+set(Matplot++_DIR "C:\\Program Files\\matplotplusplus 1.1.0\\lib\\cmake\\Matplot++")
 
-set(CMAKE_CXX_STANDARD 11)
+find_package(Torch REQUIRED)
+find_package(Matplot++ REQUIRED)
+
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${TORCH_CXX_FLAGS}")
 
 # Add source to this project's executable.
