@@ -1,4 +1,4 @@
-#include "AgentTest.h"
+ #include "AgentTest.h"
 
 #include <iostream>
 #include <fstream>
@@ -8,133 +8,129 @@
 #include <vector>
 #include <numeric>
 
+#include "../ACER/Trajectory.h"
+#include "../ACER/ACERParameters.h"
 #include "TestEnvironment.h"
 #include "../../Environment/State.h"
 
-using namespace RLGameAgent;
-
-AgentTest::AgentTest()
+AgentTest::AgentTest(bool isContinous, bool recordOutput) : 
+    isContinous(isContinous), recordOutput(recordOutput)
 {
-    // Model.
-    float lr1 = 1e-3f;
-    float lr2 = 1e-3f;
-    unsigned int nActions = 2;
-    unsigned int maxActions = 2;
-    int64_t inputDims = 4;
-    int64_t layer1Dims = 16;
-    int64_t layer2Dims = 32;
-    unsigned int memSize = 1000000;
-    float gamma = 0.99f;
-    float tau = 0.005f;
-    unsigned int batchSize = 1024;
-    int rewardScale = 2;
+    // Setup Model.
+    // ...
+    ACERParameters params;
+    params.learningRate = 1e-3f;
+    params.nActions = 1;
+    params.nPossibleActions = 4;
+    params.inputDims = 4;
+    params.hiddenLayerDims = 128;
+    params.actionLayerDims = 256;
+    params.memSize = 1000000;
+    batchSize = 16;
+    params.batchSize = batchSize;
+    params.biasWeight = 0.1f;
+    params.gamma = 0.99f;
+    params.traceMax = 10;
 
-    agent = new SACAgent(lr1, lr2, nActions, maxActions,
-        inputDims, layer1Dims, layer2Dims,
-        memSize, gamma, tau, batchSize, rewardScale);
+    agent = std::make_unique<ACERAgent>(params);
 }
 
 void AgentTest::Train()
 {
-    // Random engine.
+    // Setup Enviroment
     std::random_device rd;
     std::mt19937 re(rd());
     std::uniform_int_distribution<> dist(-5, 5);
-
-    // Environment.
+    
     float x = float(dist(re)); // goal x pos
     float y = float(dist(re)); // goal y pos
     TestEnvironment env(x, y);
 
-    // Model.
-   /* float lr1 = 1e-3f;
-    float lr2 = 1e-3f;
-    unsigned int nActions = 2;
-    unsigned int maxActions = 2; 
-    int64_t inputDims = 4;
-    int64_t layer1Dims = 16;
-    int64_t layer2Dims = 32;
-    unsigned int memSize = 1000000;
-    float gamma = 0.99f;
-    float tau = 0.005f;
-    unsigned int batchSize = 1024;
-    int rewardScale = 2;
-
-    SACAgent agent(lr1, lr2, nActions, maxActions,
-        inputDims, layer1Dims, layer2Dims,
-        memSize, gamma, tau, batchSize, rewardScale);*/
-    // TODO: He normalizes the network, unknown if relevent
+    // Setup Output.
+    std::ofstream out;
+    if (recordOutput)
+    {
+        out.open("Data/data.csv");
+        out << 1 << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
+    }
 
     // Training loop.
-    unsigned int nGames = 10000;
-    unsigned int n_steps = 2048;
-
-    // Output.
-    std::ofstream out;
-    out.open("Data/data.csv");
-
-    // episode, agent_x, agent_y, goal_x, goal_y, STATUS=(PLAYING, WON, LOST, RESETTING)
-    out << 1 << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
-
-    // Counter.
-    unsigned int counter = 0;
-
-    // Average reward.
-    // TODO: Reward object
+    unsigned int nEpisodes = 2000;
+    unsigned int maxEpisodeLength = 2048;
+    unsigned int winCount = 0;
     float bestAvgReward = -100000; 
     float avgReward = 0;
-    vector<float> rewardHistory;
-    int winCount = 0;
-
-    // REVIEW: From here down
-    for (unsigned int episode = 1; episode <= nGames; episode++)
+    std::vector<float> rewardHistory;
+    for (unsigned int episode = 1; episode <= nEpisodes; episode++)
     {
-        printf("Game %u/%u\n", episode, nGames);
+        printf("Game %u/%u\n", episode, nEpisodes);
 
         // State of Env.
-        State observation;
+        GameAgent::State observation;
         torch::Tensor state = env.State();
-        observation.AddDelta(pair<string, float>("GoalX", x));
-        observation.AddDelta(pair<string, float>("GoalY", y));
-        observation.AddDelta(pair<string, float>("PosX", state[0][0].item<float>()));
-        observation.AddDelta(pair<string, float>("PosY", state[0][1].item<float>()));
+        observation.AddDelta("GoalX", x);
+        observation.AddDelta("GoalY", y);
+        observation.AddDelta("PosX", state[0][0].item<float>());
+        observation.AddDelta("PosY", state[0][1].item<float>());
 
+        // Create trajectory
+        Trajectory trajectory(maxEpisodeLength, observation.Size(), 1, 4);
+
+        //cout << observation << endl;
+        std::vector<Trajectory> trajectories;
         float reward = 0;
-
         unsigned int i = 0;
-        for (i = 1; i <= n_steps; i++)
+        for (i = 1; i <= maxEpisodeLength; i++)
         {
+            //Trajectory trajectory(1, observation.Size(), 1, 4);
+
             // Play
-            torch::Tensor action = agent->ChooseAction(observation);
+            auto predictedAction = agent->PredictAction(observation.ToTensor());
 
-            
-            /*cout << "Action: " << endl;
-            cout << action << endl;*/
-            
+            float xAct = 0.0f;
+            float yAct = 0.0f;
+            switch(static_cast<int>(predictedAction.at(0)))
+            {
+            case 0:
+                xAct = 0.5f;
+                break;
+            case 1:
+                xAct = -0.5f;
+                break;
+            case 2:
+                yAct = 0.5f;
+                break;
+            case 3:
+                yAct = -0.5f;
+                break;
+            default:
+                std::cout << "Something went wrong" << std::endl;
+                break;
+            }
 
-            float xAct = action[0].item<float>();
-            float yAct = action[1].item<float>();
             auto sd = env.Act(xAct, yAct); // std::tuple<state, status, terminal>
 
             // Get parameters
-            State newObservation;
+            GameAgent::State newObservation;
             torch::Tensor newState = std::get<0>(sd);
-            newObservation.AddDelta(pair<string, float>("GoalX", x));
-            newObservation.AddDelta(pair<string, float>("GoalY", y));
-            newObservation.AddDelta(pair<string, float>("X", newState[0][0].item<float>()));
-            newObservation.AddDelta(pair<string, float>("Y", newState[0][1].item<float>()));
+            newObservation.AddDelta("GoalX", x);
+            newObservation.AddDelta("GoalY", y);
+            newObservation.AddDelta("X", newState[0][0].item<float>());
+            newObservation.AddDelta("Y", newState[0][1].item<float>());
 
-            std::vector<float> actions = { xAct, yAct };
             float newReward = env.Reward(std::get<1>(sd))[0].item<float>();
-            // bool terminal = (std::get<1>(sd) == TestEnvironment::STATUS::WON) ? true : false;
+            //bool terminal = (std::get<1>(sd) == TestEnvironment::STATUS::WON) ? true : false;
             bool terminal = std::get<2>(sd)[0].item<float>();
 
             // Update Agent
             reward += newReward;
-            out << episode << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << std::get<1>(sd) << "\n";
+
+            if (recordOutput)
+                out << episode << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << std::get<1>(sd) << "\n";
             
-            agent->UpdateMemory(observation, actions, newReward, newObservation, terminal);
-            agent->Learn();
+            agent->UpdateMemory(observation, predictedAction, newReward,
+                                newObservation, terminal);
+
             observation = newObservation;
 
             // Check game status
@@ -148,14 +144,16 @@ void AgentTest::Train()
         }
 
         // Update agent
-        // agent->Learn();
+        //agent->Learn();
 
         // Reset game
         x = float(dist(re));
         y = float(dist(re));
         env.SetGoal(x, y);
         env.Reset();
-        out << episode << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
+
+        if (recordOutput)
+            out << episode << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
 
         // Get rewards
         rewardHistory.push_back(reward);
@@ -165,18 +163,18 @@ void AgentTest::Train()
         else
             avgReward = accumulate(rewardHistory.begin(), rewardHistory.end(), 0.0) / rewardHistory.size();
         
-        cout << "Step count: " << i << endl;
-        printf("Game's Won %u/%u\n", winCount, nGames);
-        cout << "Reward: " << reward << endl;
-        cout << "AvgReward: " << avgReward << endl;
+        std::cout << "Step count: " << i << std::endl;
+        printf("Game's Won %u/%u\n", winCount, nEpisodes);
+        std::cout << "Reward: " << reward << std::endl;
+        std::cout << "AvgReward: " << avgReward << std::endl;
 
-        if ((avgReward > bestAvgReward && rewardHistory.size() >= 100) || episode % 10000 == 0)
+        if ((avgReward > bestAvgReward && rewardHistory.size() >= 100))
         {
             bestAvgReward = avgReward;
             agent->SaveModel();
         }
 
-        cout << "BestAvgReward: " << bestAvgReward << endl << endl;
+        std::cout << "BestAvgReward: " << bestAvgReward << std::endl << std::endl;
     }
 
     out.close();
@@ -184,92 +182,95 @@ void AgentTest::Train()
 
 void AgentTest::Test()
 {
-    // Random engine.
+    // Test Code
+    // ...
+    // Setup Enviroment
     std::random_device rd;
     std::mt19937 re(rd());
     std::uniform_int_distribution<> dist(-5, 5);
 
-    // Environment.
     float x = float(dist(re)); // goal x pos
     float y = float(dist(re)); // goal y pos
     TestEnvironment env(x, y);
 
-    // Model.
-    /*float lr1 = 1e-3f;
-    float lr2 = 1e-3f;
-    unsigned int nActions = 2;
-    unsigned int maxActions = 2;
-    int64_t inputDims = 4;
-    int64_t layer1Dims = 256;
-    int64_t layer2Dims = 256;
-    unsigned int memSize = 1000000;
-    float gamma = 0.99f;
-    float tau = 0.005f;
-    unsigned int batchSize = 1024;
-    int rewardScale = 2;
+    // Setup Output.
+    std::ofstream out;
+    if (recordOutput)
+    {
+        out.open("Data/data.csv");
+        out << 1 << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
+    }
 
-    SACAgent agent(lr1, lr2, nActions, maxActions,
-        inputDims, layer1Dims, layer2Dims,
-        memSize, gamma, tau, batchSize, rewardScale);*/
-
-    // Load Model
-    // agent.LoadModel();
+    // Load Agent
+    agent->LoadModel();
 
     // Training loop.
-    unsigned int nGames = 10000;
-    unsigned int nSteps = 2048;
-
-    // Output.
-    std::ofstream out;
-    out.open("Data/data.csv");
-
-    // episode, agent_x, agent_y, goal_x, goal_y, STATUS=(PLAYING, WON, LOST, RESETTING)
-    out << 1 << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
-
-    // Counter.
+    unsigned int nEpisodes = 1000;
+    unsigned int maxEpisodeLength = 2048;
     unsigned int winCount = 0;
-
-    for (unsigned int episode = 1; episode <= nGames; episode++)
+    for (unsigned int episode = 1; episode <= nEpisodes; episode++)
     {
-        printf("Game %u/%u\n", episode, nGames);
+        printf("Game %u/%u\n", episode, nEpisodes);
 
         // State of Env.
-        State observation;
+        GameAgent::State observation;
         torch::Tensor state = env.State();
-        observation.AddDelta(pair<string, float>("GoalX", x));
-        observation.AddDelta(pair<string, float>("GoalY", y));
-        observation.AddDelta(pair<string, float>("PosX", state[0][0].item<float>()));
-        observation.AddDelta(pair<string, float>("PosY", state[0][1].item<float>()));
+        observation.AddDelta("GoalX", x);
+        observation.AddDelta("GoalY", y);
+        observation.AddDelta("PosX", state[0][0].item<float>());
+        observation.AddDelta("PosY", state[0][1].item<float>());
 
         float reward = 0;
-
-        for (unsigned int i = 0; i < nSteps; i++)
+        unsigned int i = 0;
+        for (i = 1; i <= maxEpisodeLength; i++)
         {
             // Play
-            torch::Tensor action = agent->ChooseAction(observation);
-            float xAct = action[0].item<float>();
-            float yAct = action[1].item<float>();
+            auto predictedAction = agent->PredictAction(observation.ToTensor());
+
+            float xAct = 0.0f;
+            float yAct = 0.0f;
+            switch (static_cast<int>(predictedAction.at(0)))
+            {
+            case 0:
+                xAct = 0.5f;
+                break;
+            case 1:
+                xAct = -0.5f;
+                break;
+            case 2:
+                yAct = 0.5f;
+                break;
+            case 3:
+                yAct = -0.5f;
+                break;
+            default:
+                std::cout << "Something went wrong" << std::endl;
+                break;
+            }
+
             auto sd = env.Act(xAct, yAct); // std::tuple<state, status, terminal>
 
             // Get parameters
-            State newObservation;
+            GameAgent::State newObservation;
             torch::Tensor newState = std::get<0>(sd);
-            newObservation.AddDelta(pair<string, float>("GoalX", x));
-            newObservation.AddDelta(pair<string, float>("GoalY", y));
-            newObservation.AddDelta(pair<string, float>("X", newState[0][0].item<float>()));
-            newObservation.AddDelta(pair<string, float>("Y", newState[0][1].item<float>()));
-
-            float newReward = env.Reward(std::get<1>(sd))[0].item<float>();
-            bool terminal = (std::get<1>(sd) == TestEnvironment::STATUS::WON) ? true : false;
+            newObservation.AddDelta("GoalX", x);
+            newObservation.AddDelta("GoalY", y);
+            newObservation.AddDelta("X", newState[0][0].item<float>());
+            newObservation.AddDelta("Y", newState[0][1].item<float>());
 
             // Update Agent
+            float newReward = env.Reward(std::get<1>(sd))[0].item<float>();
             reward += newReward;
-            out << 1 << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << std::get<1>(sd) << "\n";
+
+            if (recordOutput)
+                out << episode << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << std::get<1>(sd) << "\n";
+
+            observation = newObservation;
 
             // Check game status
             if (std::get<2>(sd)[0].item<float>())
             {
-                if (terminal)
+                if (std::get<1>(sd) == TestEnvironment::STATUS::WON)
                     winCount++;
 
                 break;
@@ -281,11 +282,12 @@ void AgentTest::Test()
         y = float(dist(re));
         env.SetGoal(x, y);
         env.Reset();
-        out << episode << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
 
-        // Print results
-        printf("Game's Won %u/%u\n", winCount, nGames);
-        cout << endl;
+        if (recordOutput)
+            out << episode << ", " << env.pos(0) << ", " << env.pos(1) << ", " << env.goal(0) << ", " << env.goal(1) << ", " << env.RESETTING << "\n";
+
+        std::cout << "Step count: " << i << std::endl;
+        printf("Game's Won %u/%u\n", winCount, nEpisodes);
     }
 
     out.close();
